+++ src/main/java/utils/Reader.java
@@ -70,7 +70,7 @@ public class Reader {
 
                     String originS = eElement.getElementsByTagName("from").item(0).getTextContent();
                     String destinyS = eElement.getElementsByTagName("to").item(0).getTextContent();
-                    double distance = Double.parseDouble(eElement.getElementsByTagName("distance").item(0).getTextContent());
+                    int distance = Integer.parseInt(eElement.getElementsByTagName("distance").item(0).getTextContent());
                     String name = eElement.getElementsByTagName("name").item(0).getTextContent();
                     
                     Point origin = Utils.getPointByName(originS,points);
